PROJ-48 feat: add utility function handleRecoveryAndEmailLinkErrors
