+++ features/index.ts
@@ -1,6 +1,7 @@
 export * from './ConfirmedEmail'
 export * from './ExpiredEmailLink'
 export * from './ForgotPassword'
+export * from './PasswordRecovery'
 export * from './appAlert'
 export * from './newPassword'
 export * from './signUp'
