+++ shared/lib/index.ts
@@ -1,4 +1,5 @@
 export * from './checkTokenValidity'
 export * from './handleNetworkError'
+export * from './handleRecoveryAndEmailLinkErrors'
 export * from './handleServerError'
 export * from './nextnprogress'
